requirements.txt
Django==4.2
djangorestframework==3.16.1
PyJWT==2.10.1
python-dotenv==1.2.1
